requirements.txt
google-api-core
googleapis-common-protos
grpcio
grpcio-reflection
grpcio-tools
ondewo-client-utils>=3.1.0
polling
protobuf
regex
